Move map to its own class
