+++ SimplyTransport/controllers/root.py
@@ -1,6 +1,7 @@
 from litestar import Controller, Response, get
 from litestar.response import Template
 from litestar.di import Provide
+from SimplyTransport.domain.maps.maps import Map
 
 from SimplyTransport.domain.maps.polylines import PolyLineColors, RoutePolyLine
 
@@ -23,6 +24,7 @@ class RootController(Controller):
     dependencies = {
         "event_repo": Provide(provide_event_repo),
         "map_service": Provide(MapService, sync_to_thread=False),
+        "map": Provide(Map, sync_to_thread=False),
     }
 
     @get("/")
@@ -82,25 +84,26 @@ class RootController(Controller):
         return Template("gtfs_search/route_search.html")
 
     @get("/test/map")
-    async def map(self, map_service: MapService) -> Template:
-        map_service.setup_defaults()
+    async def map(self, map: Map) -> Template:
+        map = Map()
+        map.setup_defaults()
         marker1 = StopMarker("Test Stop", "123", "TST", 53.0, -7.0, color=MarkerColors.RED)
-        marker1.add_to(map_service.map, type_of_marker="circle")
+        marker1.add_to(map.map, type_of_marker="circle")
 
         marker2 = StopMarker("Test Stop 2", "124", "TST2", 53.1, -7.1, create_links=False)
-        marker2.add_to(map_service.map, type_of_marker="regular")
+        marker2.add_to(map.map, type_of_marker="regular")
 
         route1 = RoutePolyLine(
             "1", "Test Route", "Test Operator", [(53.0, -7.0), (53.1, -7.1)], route_color=PolyLineColors.BLUE
         )
-        route1.add_to(map_service.map)
+        route1.add_to(map.map)
 
         bus1 = BusMarker("Test Bus", "1", "Test Operator", 53.2, -7.2, color=MarkerColors.GREEN)
-        bus1.add_to(map_service.map)
+        bus1.add_to(map.map)
 
         return Template(
             template_name="map.html",
             context={
-                "map": map_service.render(),
+                "map": map.render(),
             },
         )
